requirements.txt
asgiref==3.2.10
astroid==2.4.2
cffi==1.14.1
cryptography==3.0
dj-database-url==0.5.0
Django==3.1
django-widget-tweaks==1.4.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pycparser==2.20
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
appdirs==1.4.2
django-bootstrap4==2.2.0
django-bootstrap-datepicker-plus==3.0.5
django-datetimepicker==3.14
django-cleanup==5.1.0
django-tables2==2.3.2
django-filter==2.4.0
Markdown==3.3.3
djangorestframework==3.5.4
djangorestframework-jsonap==3.2.0

djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
django


